auth_service/requirements.txt
fastapi
python-dotenv
requests_oauthlib
uvicorn
